Export the public OpenPGP key and sample sources.list
ti-debpkgs.gpg.key: The public part of the OpenPGP key is required for
adding the repo to sources.list.
ti-debpkgs.sources: This is a sample sources.list which can be used to
install packages from bookworm / stable.

Update README to mention how to use the PPA using the sources file.

Signed-off-by: Sai Sree Kartheek Adivi <s-adivi@ti.com>